+++ shuup/utils/pdf.py
